Documenting all types
