+++ OpenCANMonitor/Can Library/Swift Wrapper/Enumerations/CANStatus.swift
@@ -1,5 +1,5 @@
 //
-//  PCANError.swift
+//  CANStatus.swift
 //  CanMonitor
 //
 //  Created by Taylor Lineman on 9/5/23.
@@ -7,37 +7,68 @@
 
 import Foundation
 
-enum PCANStatus: CaseIterable, LocalizedError {
+/// A CAN status code that is returned by almost all **PCBUSB Library** functions
+enum CANStatus: CaseIterable, LocalizedError {
+    /// No error.
     case ok
+    /// Transmit buffer in CAN controller is full.
     case xmtfull
+    /// CAN controller was read too late.
     case overrun
+    /// Bus error: an error counter reached the 'light' limit.
     case buslight
+    /// Bus error: an error counter reached the 'heavy' limit.
     case busheavy
+    /// Bus error: an error counter reached the 'warning' limit.
     case buswarning
+    /// Bus error: the CAN controller is error passive.
     case buspassive
+    /// Bus error: the CAN controller is in bus-off state.
     case busoff
+    /// Mask for all bus errors.
     case anybuserr
+    /// Receive queue is empty.
     case qrcvempty
+    /// Receive queue was read too late.
     case qoverrun
+    /// Transmit queue is full.
     case qxmtfull
+    /// Test of the CAN controller hardware registers failed (no hardware found).
     case regtest
+    /// Driver not loaded. (`libPCBUSB.dyld` did not load correctly).
     case nodriver
+    /// Hardware already in use by a Net.
     case hwinuse
+    /// A Client is already connected to the Net.
     case netinuse
+    /// Hardware handle is invalid.
     case illhw
+    ///Net handle is invalid.
     case illnet
+    /// Client handle is invalid.
     case illclient
+    /// Mask for all handle errors.
     case illhandle
+    /// Resource (FIFO, Client, timeout) cannot be created.
     case resource
+    /// Invalid parameter.
     case illparamtype
+    /// Invalid parameter value.
     case illparamval
+    /// Unknown error.
     case unknown
+    /// Invalid data, function, or action
     case illdata
+    /// Driver object state is wrong for the attempted operation.
     case illmode
+    /// An operation was successfully carried out, however, irregularities were registered.
     case caution
+    /// Channel is not initialized.
     case initialize
+    /// Invalid operation.
     case illoperation
     
+    /// The description of the error. Originally sourced from `PCBUSB.h`
     var errorDescription: String? {
         switch self {
         case .ok:
@@ -67,7 +98,7 @@ enum PCANStatus: CaseIterable, LocalizedError {
         case .regtest:
             return "Test of the CAN controller hardware registers failed (no hardware found)"
         case .nodriver:
-            return "Driver not loaded"
+            return "Driver not loaded. ('libPCBUSB.dyld' did not load correctly)"
         case .hwinuse:
             return "Hardware already in use by a Net"
         case .netinuse:
@@ -95,16 +126,18 @@ enum PCANStatus: CaseIterable, LocalizedError {
         case .caution:
             return "An operation was successfully carried out, however, irregularities were registered"
         case .initialize:
-            return "Channel is not initialized [Value was changed from 0x40000 to 0x4000000]"
+            return "Channel is not initialized"
         case .illoperation:
-            return "Invalid operation [Value was changed from 0x80000 to 0x8000000]"
+            return "Invalid operation"
         }
     }
 }
 
-extension PCANStatus: RawRepresentable {
+extension CANStatus: RawRepresentable {
     typealias RawValue = UInt32
     
+    /// An optional RawRepresentable initializer that initializes a status from its ``RawValue-swift.typealias``
+    /// - Parameter rawValue: The rawValue to initialize with. Should be the UInt32 value that is found in the `PCBUSB.h` definitions.
     init?(rawValue: UInt32) {
         for element in Self.allCases {
             if element.rawValue == rawValue {
@@ -115,6 +148,7 @@ extension PCANStatus: RawRepresentable {
         return nil
     }
     
+    /// The RawValue of the status. These values link back to the original C values found in `PCBUSB.h`
     var rawValue: UInt32 {
         switch self {
         case .ok:
