+++ OpenCANMonitor/Settings/GeneralSettingsView.swift
@@ -14,30 +14,30 @@ struct GeneralSettingsView: View {
     @AppStorage("defaultBus") var defaultBus: Int = 0
     @AppStorage("defaultBaud") var defaultBaud: Int = 0
     
-    @State var usbBus: PCANUSBBus = .bus1
-    @State var baudRate: PCANBaudRate = .baud1M
+    @State var usbBus: USBBus = .bus1
+    @State var baudRate: BaudRate = .baud1M
     
     var body: some View {
         Form {
             Picker("Default Interface", selection: $usbBus) {
                 // Loop over all the interfaces, but drop the last because it is the none interface.
-                ForEach(PCANUSBBus.allCases.dropLast()) { interface in
+                ForEach(USBBus.allCases.dropLast()) { interface in
                     Text(interface.displayName)
                         .tag(interface)
                 }
             }
             .onChange(of: usbBus) { newValue in
-                defaultBus = PCANUSBBus.allCases.firstIndex(of: newValue)!
+                defaultBus = USBBus.allCases.firstIndex(of: newValue)!
             }
             
             Picker("Default Baud Rate", selection: $baudRate) {
-                ForEach(PCANBaudRate.allCases) { rate in
+                ForEach(BaudRate.allCases) { rate in
                     Text(rate.displayName)
                         .tag(rate)
                 }
             }
             .onChange(of: baudRate) { newValue in
-                defaultBaud = PCANBaudRate.allCases.firstIndex(of: newValue)!
+                defaultBaud = BaudRate.allCases.firstIndex(of: newValue)!
             }
         }
         .onAppear {
@@ -46,8 +46,8 @@ struct GeneralSettingsView: View {
             let defaultBusIndex = UserDefaults.standard.integer(forKey: "defaultBus")
             let defaultBaudIndex = UserDefaults.standard.integer(forKey: "defaultBaud")
             
-            usbBus = PCANUSBBus.allCases[defaultBusIndex]
-            baudRate = PCANBaudRate.allCases[defaultBaudIndex]
+            usbBus = USBBus.allCases[defaultBusIndex]
+            baudRate = BaudRate.allCases[defaultBaudIndex]
         }
     }
 }
